add ons-boundary script + updates

+++ src/scripts/extract.py
@@ -1,4 +1,6 @@
+import os
 from io import BytesIO
+from pathlib import Path
 from typing import Annotated
 from zipfile import ZipFile
 
@@ -11,6 +13,7 @@ from safer_streets_core.database import duckdb_spatial_connector
 from safer_streets_core.utils import data_dir, latest_month
 
 # This script extracts street-level crime data from zipped CSV files and saves them as Parquet files.
+# or creates a duckdb table
 # archives can be downloaded from https://data.police.uk/data/ (see also download_archive in utils.py)
 
 OUT_PATH = data_dir() / "extracted"
@@ -73,36 +76,40 @@ def latest(*, keep_existing: bool = False) -> None:
 
 
 @app.command()
-def to_database() -> None:
-    # extract the latest archive and overwrite any existing data
-    zipfile = "https://data.police.uk/data/archive/latest.zip"
-    print(f"Downloading {zipfile}...")
-    response = requests.get(zipfile, stream=True)
-    file_size = int(response.headers.get("content-length", 0))
+def to_database(force_download: bool = False) -> None:
 
-    with open("/tmp/latest.zip", "wb") as fd, tqdm(total=file_size, unit="B", unit_scale=True) as progress_bar:
-        for data in response.iter_content(1024**2):
-            progress_bar.update(len(data))
-            fd.write(data)
+    tmpfile = Path("/tmp/latest.zip")
+    if force_download or not tmpfile.exists():
+        # extract the latest archive and overwrite any existing data
+        zipfile = "https://data.police.uk/data/archive/latest.zip"
+        print(f"Downloading {zipfile}...")
+        response = requests.get(zipfile, stream=True)
+        file_size = int(response.headers.get("content-length", 0))
+
+        with open("/tmp/latest.zip", "wb") as fd, tqdm(total=file_size, unit="B", unit_scale=True) as progress_bar:
+            for data in response.iter_content(1024**2):
+                progress_bar.update(len(data))
+                fd.write(data)
+    else:
+        print(f"Using cached {tmpfile}...")
 
     print("Creating table crime_data...")
-    con = duckdb_spatial_connector(data_dir() / "safer_streets.db")
+    con = duckdb_spatial_connector(data_dir() / os.environ["SAFER_STREETS_CRIME_DATABASE"])
 
     # limited support for **/ glob, but ????-?? is a reasonable workaround
     con.execute("""
     DROP TABLE IF EXISTS crime_data;
     CREATE TABLE crime_data AS
-    SELECT * FROM read_csv('zip://../data/police_uk_crime_data_latest.zip/????-??/*-street.csv', normalize_names = true);
-    ALTER TABLE crime_data ADD COLUMN geometry GEOMETRY;
+    SELECT * FROM read_csv('zip:///tmp/latest.zip/????-??/*-street.csv', normalize_names = true);
+    ALTER TABLE crime_data ADD COLUMN geom GEOMETRY;
     UPDATE crime_data
-    SET geometry =
-        ST_Transform(
-            ST_Point(Longitude, Latitude),
+    SET geom = ST_Transform(
+            ST_Point(longitude, latitude),
             'EPSG:4326',
             'EPSG:27700',
             always_xy := true
         )
-    WHERE latitude IS NOT NULL AND longitude IS NOT NULL;
+    -- WHERE latitude IS NOT NULL AND longitude IS NOT NULL;
     """)
